refactor: simplify the theming system (#34)
Rework the theming system by simplifying the `UAppTheme` global.  
Instead of having to create whole new `AppTheme` structs, you can now
create new specific style structs and add them to the global theme from your
code.

**Remove the theming properties from the widgets:**  
This choice was made due to limitations with the current theming system.
Lets take the `UButton` widget as an example.
You could create custom variants using this widget as the base, but
couldn't use them in more complex widgets such as the `UModal`.
Hence I decided to make it so this library has a more locked in style
(that can still be tuned a little bit with the global theme).